shortened the excessive 4byte placeholder to 2

+++ src/db/alloc.h
@@ -12,7 +12,7 @@
 int8_t *db_char_copy(int8_t *);
 Schema *db_schema_alloc();
 Attribute *db_attribute_alloc();
-Attribute **db_attributep_alloc(uint16_t);
+Attribute **db_attributep_alloc(uint8_t);
 Query *db_query_alloc();
 
 #endif
